Move RDS querying/writing to its own file
It was merged in with scribe before for no real reason, this splits it out into its own module to make more sense and simplify usage (`scribe.py` is mostly internal-only, whereas `rds.py` is almost all user-facing functions)
