+++ tools/stats/print_test_stats.py
@@ -20,7 +20,8 @@ from typing_extensions import TypedDict
 from tools.stats.s3_stat_parser import (newify_case, get_S3_object_from_bucket, get_test_stats_summaries_for_job,
                                         Report, Status, Commit, HAVE_BOTO3, Version2Case, VersionedReport,
                                         Version1Report, Version2Report, ReportMetaMeta)
-from tools.stats.scribe import send_to_scribe, rds_write, register_rds_schema, schema_from_sample
+from tools.stats.scribe import send_to_scribe
+from tools.stats.rds import rds_write, register_rds_schema, schema_from_sample
 
 
 SimplerSuite = Dict[str, Version2Case]
